fix: harden pack subscription security and fix double query
Validate pack IDs against accessible packs before storing in session,
add UUID validation before all get_object_or_404 calls, and evaluate
selected_packs queryset to list in the view to avoid double query in
template. Move pack_fighter/pack_rule fixtures to shared conftest.
