+++ gyrinx/core/templates/core/list_new.html
@@ -7,7 +7,7 @@
     {% include "core/includes/back.html" %}
     <div class="col-12 col-md-8 col-lg-6 px-0 vstack gap-3">
         <h1 class="h3">Create a new List</h1>
-        {% if selected_packs.exists %}
+        {% if selected_packs %}
             <div class="border rounded p-3">
                 <div class="d-flex justify-content-between align-items-center mb-2">
                     <span class="fw-medium">Content Packs</span>
